+++ app/lib/util-service.js
@@ -612,7 +612,6 @@ angular.module('lizard-nxt')
    */
   this.union = _.union;
 
-
   /*
    * @function
    * @description - return the average rain per square kilometer, given:
@@ -643,4 +642,34 @@ angular.module('lizard-nxt')
       }
     }
   };
+
+
+  // Add comparator to sort lists on multiple properties to D3.
+  (function () {
+    d3.comparator = function () {
+      var cmps = [], accessors = [];
+
+      var comparator = function (a, b) {
+        var i = -1,
+            n = cmps.length,
+            result;
+        while (++i < n) {
+          result = cmps[i](accessors[i](a), accessors[i](b));
+          if (result !== 0) return result;
+        }
+        return 0;
+      };
+
+      comparator.order = function (cmp, accessor) {
+        cmps.push(cmp);
+        accessors.push(accessor || identity);
+        return comparator;
+      };
+
+      return comparator;
+    };
+
+    function identity(d) { return d; }
+  })();
+
 }]);