+++ app/lizard-nxt.js
@@ -21,10 +21,7 @@ if (window.RavenEnvironment) {
 angular.module("lizard-nxt", [
   'omnibox',
   'restangular',
-  'ngSanitize',
-  'ngCsv',
-  'ui.bootstrap',
-  'ui.utils'
+  'global-state'
 ])
 // Decorator for ngExceptionHandler to log exceptions to sentry
   .config(function ($provide) {
@@ -160,6 +157,8 @@ angular.module('lizard-nxt')
    'ClickFeedbackService',
    'user',
    'versioning',
+   'State',
+   'MapService',
 
   function ($scope,
             $http,
@@ -172,150 +171,12 @@ angular.module('lizard-nxt')
             TimeseriesService,
             ClickFeedbackService,
             user,
-            versioning) {
+            versioning,
+            State,
+            MapService) {
 
   $scope.user = user;
   $scope.versioning = versioning;
-
-  var MIN_TIME_FOR_EXTENT = (new Date(2010, 0, 0, 0, 0, 0, 0)).getTime();
-  var MAX_TIME_FOR_EXTENT = (new Date(2015, 0, 0, 0, 0, 0, 0)).getTime();
-
-  // BOX MODEL
-  /**
-   * @memberOf angular.module('lizard-nxt')
-  .MasterCtrl
-   * @summary Box model
-   *
-   * @description Box model holds properties to render the omnibox.
-   *
-   * @property {object} box - Box model
-   * @property {boolean} box.detailMode - Detail mode, defaults to false.
-   */
-  $scope.box = {
-    contextSwitchMode: false, // Switch between card or fullscreen
-    query: null, // Search bar query
-    showCards: false,// Only used for search results
-    type: 'point', // Default box type
-    //type: undefined, // Should this be set via the hashGetterSetter????
-    content: {}, // Inconsistently used to store data to display in box
-    changed: Date.now(),
-    mouseLoc: [] // Used to draw 'bolletje' on elevation profile
-  };
-  // BOX MODEL
-
-  // TOOLS
-  $scope.tools = {
-    active: 'point', //NOTE: make list?
-  };
-
-  $scope.tooltips = CabinetService.tooltips;
-
-  /**
-   * @function
-   * @memberOf angular.module('lizard-nxt')
-  .MasterCtrl
-   *
-   * @summary Toggle tool from "name" to "none".
-   *
-   * @desc Sets tool.active model on scope to name of the tool if tool disabled
-   * or "none" if tool is already enabled.
-   *
-   * @param {string} name name of the tool to toggle
-   *
-   */
-  $scope.toggleTool = function (name) {
-
-    if (["point", "line", "area"].indexOf(name) > -1) {
-      $scope.box.type = name;
-    } else {
-      throw new Error("Attemped to assign an illegal value ('"
-        + name
-        + "') to $scope.box.type. Only 'point', 'line' and 'area' are accepted values."
-      );
-    }
-  };
-
-  /**
-   * Switch between contexts.
-   *
-   * @param {string} context - Context name to switch to
-   */
-  //$scope.switchContext = function (context) {
-    //$scope.box.context = context;
-  //};
-
-  // MAP MODEL is set by the map-directive
-  $scope.mapState = {};
-
-  // TODO: check what this does
-  $scope.$watch('mapState.here', function (n, o) {
-    if (n === o) { return true; }
-
-    var fn = function () {
-      if ($scope.box.type === 'point') {
-        $scope.box.type = 'point';
-        $scope.$broadcast('updatepoint');
-      }
-    };
-
-    if (!$scope.$$phase) {
-      $scope.$apply(fn);
-    } else {
-      fn();
-    }
-  });
-
-  // TIME MODEL
-  var now = Date.now(),
-      hour = 60 * 60 * 1000,
-      day = 24 * hour;
-
-  $scope.timeState = {
-    start: now - 6 * day,
-    end: now + day,
-    at: Math.round(now - 2.5 * day),
-    changedZoom: Date.now(),
-    zoomEnded: null,
-    aggWindow: 1000 * 60 * 5,
-    animation: {
-      playing: false,
-      enabled: false,
-    }
-  };
-
-  $scope.timeState.aggWindow = UtilService.getAggWindow(
-    $scope.timeState.start,
-    $scope.timeState.end,
-    window.innerWidth
-  );
-  // END TIME MODEL
-
-  /**
-   * Watch to restrict values of timeState.
-   */
-  $scope.$watch('timeState.changedZoom', function (n, o) {
-    if (n === o || $scope.timeState.changeOrigin === 'master') { return true; }
-    if ($scope.timeState.start < MIN_TIME_FOR_EXTENT) {
-      $scope.timeState.changeOrigin = 'master';
-      $scope.timeState.start = MIN_TIME_FOR_EXTENT;
-    }
-    if ($scope.timeState.end > MAX_TIME_FOR_EXTENT) {
-      $scope.timeState.changeOrigin = 'master';
-      $scope.timeState.end = MAX_TIME_FOR_EXTENT;
-    }
-  });
-
-  $scope.$watch('box.type', function (n, o) {
-    UtilService.addNewStyle(
-      "#map * {cursor:" + (n === "line" ? "crosshair" : "") + ";}"
-    );
-  });
-
-  //TODO: move to raster-service ?
-
-  $scope.raster = {
-    changed: Date.now()
-  };
 
   // KEYPRESS
 
@@ -329,32 +190,25 @@ angular.module('lizard-nxt')
       return;
     }
 
+    // pressed ESC
     if ($event.which === 27) {
-      // If detailMode is active, close that
-      if ($scope.box.contextSwitchMode) {
-        // this shouldn't matter until dates gte 01-12-2014
-        $scope.box.contextSwitchMode = false;
-      } else {
-        // Or else, reset the omnibox AND searchbar state:
 
-        $scope.box.type = "point";
-        $scope.box.content = {};
-        $scope.mapState.here = undefined;
-        $scope.mapState.points = [];
-        ClickFeedbackService.emptyClickLayer($scope.mapState);
+      State.box.type = "point";
+      State.spatial.here = undefined;
+      State.spatial.points = [];
+      ClickFeedbackService.emptyClickLayer(MapService);
 
-        // This does NOT work, thnx to Angular scoping:
-        // $scope.geoquery = "";
-        //
-        // ...circumventing-angular-weirdness teknique:
-        document.querySelector("#searchboxinput").value = "";
-      }
+      // This does NOT work, thnx to Angular scoping:
+      // $scope.geoquery = "";
+      //
+      // ...circumventing-angular-weirdness teknique:
+      document.querySelector("#searchboxinput").value = "";
     }
-
-    // play pause timeline with Space.
-    if ($event.which === 32) {
-      $scope.timeState.playPauseAnimation();
-    }
+    // TODO: move
+    // // play pause timeline with Space.
+    // if ($event.which === 32) {
+    //   $scope.timeState.playPauseAnimation();
+    // }
 
   };