+++ CHANGES.rst
@@ -1,8 +1,26 @@
 Changelog of lizard-nxt client
 ==============================
 
-Unreleased ()
--------------
+Unreleased (1.2.dev) (XXXX-XX-XX)
+---------------------------------
+
+- Removed unused images.
+
+-
+
+
+Release 1.1.0 (2014-12-17)
+--------------------------
+
+- Force cursor behaviour in point, line and area mode.
+
+- Clean up dependencies.
+
+- Store global state in seperate module.
+
+- Split map from data.
+
+- Update Angular coding guidelines.
 
 - Rain aggregation: gebiedsgemiddelde neerslag in omnibox
 
@@ -35,6 +53,13 @@ Unreleased ()
 - CSS is now preprocessed with SASS, fmbo of structure and clarity.
 
 - Minimize cards based on screen size and size of cards.
+
+
+Release 1.0.2 (2014-12-16)
+--------------------------
+
+- Help button
+
 
 Release 1.0.0 (2014-12-01)
 ---------------------------